+++ R/BulkLoad.R
@@ -62,6 +62,14 @@ checkBulkLoadCredentials <- function(connection) {
       return(FALSE)
     }
     return(TRUE)
+  } else if (connection@dbms == "spark") {
+    if (Sys.getenv("DATABRICKS_DBFS_PATH") == "" |
+        Sys.getenv("DATABRICKS_ROOT_FOLDER") == "" |
+        Sys.getenv("DATABRICKS_STAGING_SCHEMA") == "") {
+      inform("Please set environment variables DATABRICKS_DBFS_PATH (pointing to Databricks CLI), DATABRICKS_ROOT_FOLDER, and DATABRICKS_STAGING_SCHEMA.")
+      return(FALSE)
+    }
+    return(TRUE)
   } else {
     return(FALSE)
   }
@@ -191,6 +199,64 @@ bulkLoadRedshift <- function(connection, sqlTableName, data) {
   })
   delta <- Sys.time() - start
   inform(paste("Bulk load to Redshift took", signif(delta, 3), attr(delta, "units")))
+}
+
+bulkLoadSpark <- function(connection, sqlTableName, sqlFieldNames, sqlDataTypes, data) {
+  uploadToDbfs <- function(rootFolder,
+                           fileName) {
+    command <- sprintf("%s cp %s.txt dbfs:/%s/%s.txt", 
+                       Sys.getenv("DATABRICKS_DBFS_PATH"),
+                       fileName, rootFolder, basename(fileName))
+    print(command)
+    tryCatch({
+      system(command,
+             intern = FALSE,
+             ignore.stdout = FALSE,
+             ignore.stderr = FALSE,
+             wait = TRUE,
+             input = NULL)
+    }, error = function(e) {
+      writeLines(sprintf("DBFS Upload ERROR: %s",
+                         basename(fileName)))
+    })
+  }
+  start <- Sys.time()
+  tableName <- (strsplit(x = sqlTableName, split = ".", fixed = TRUE))[[1]][[2]]
+  
+  fileName <- file.path(tempdir(), sprintf("databricks_insert_%s", uuid::UUIDgenerate(use.time = TRUE)))
+  write.table(x = data, file = sprintf("%s.txt", fileName), row.names = FALSE, col.names = FALSE, sep = "\t", quote = FALSE)
+  uploadToDbfs(rootFolder = Sys.getenv("DATABRICKS_ROOT_FOLDER"),
+               fileName = fileName)
+  
+  tableDdl <- lapply(names(sqlDataTypes), function(s) {
+    sprintf("%s %s", s, sqlDataTypes[[s]])
+  })
+  
+  loadSql <- SqlRender::loadRenderTranslateSql(sqlFilename = "databricksLoad.sql",
+                                            packageName = "DatabaseConnector",
+                                            dbms = "sql server",
+                                            tableDdl = SqlRender::translate(paste(tableDdl, collapse = ","), "spark"),
+                                            stagingDatabaseSchema = Sys.getenv("DATABRICKS_STAGING_SCHEMA"),
+                                            tableName = tableName,
+                                            rootFolder = Sys.getenv("DATABRICKS_ROOT_FOLDER"),
+                                            fileName = basename(fileName))
+  
+  finalSql <- SqlRender::loadRenderTranslateSql(sqlFilename = "databricksFinalTable.sql",
+                                            packageName = "DatabaseConnector",
+                                            dbms = "spark",
+                                            qname = sqlTableName,
+                                            stagingDatabaseSchema = Sys.getenv("DATABRICKS_STAGING_SCHEMA"),
+                                            tableName = tableName)
+  tryCatch({
+    DatabaseConnector::executeSql(connection = connection, sql = loadSql, reportOverallTime = FALSE)
+    DatabaseConnector::executeSql(connection = connection, sql = finalSql, reportOverallTime = FALSE)
+    delta <- Sys.time() - start
+    writeLines(paste("Bulk load to Databricks took", signif(delta, 3), attr(delta, "units")))
+  }, error = function(e) {
+    stop("Error in Databricks bulk upload.")
+  }, finally = {
+    try(file.remove(sprintf("%s.txt", fileName)), silent = TRUE)
+  })
 }
 
 bulkLoadHive <- function(connection, sqlTableName, sqlFieldNames, data) {